+++ rest_framework_simple_api_key/models.py
@@ -4,7 +4,7 @@ from datetime import timedelta, datetime
 from django.conf import settings
 from django.db import models
 
-from rest_framework_simple_api_key.crypto import ApiKeyCrypto
+from rest_framework_simple_api_key.crypto import get_crypto
 from rest_framework_simple_api_key.settings import package_settings
 
 
@@ -13,14 +13,12 @@ def _expiry_date():
 
 
 class AbstractAPIKeyManager(models.Manager):
-    key_crypto = ApiKeyCrypto()
-
     def get_api_key(self, pk: int | str):
         return self.get(revoked=False, pk=pk)
 
     def assign_api_key(self, obj) -> str:
         payload = {"_pk": obj.pk, "_exp": obj.expiry_date.timestamp()}
-        key = self.key_crypto.generate(payload)
+        key = get_crypto().generate(payload)
 
         return key