+++ rest_framework_simple_api_key/rotation/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class RotationConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "rest_framework_simple_api_key.rotation"
+    label = "rest_framework_simple_api_key_rotation"