+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Changelog
 
 - Migrate to Django 4.2 LTS (#46)
 - Key rotation (#42)
-- Add an exemple project showcasing how to use the package (#45)
+- Add an example project showcasing how to use the package (#45)
 
 [v0.1.2] - 2023-04-21
 ------------------